make sure stopAt will end with separator

+++ block/local/adapter.go
@@ -152,7 +152,13 @@ func (l *Adapter) Remove(obj block.ObjectPointer) error {
 }
 
 func removeEmptyDirUntil(dir string, stopAt string) {
-	for stopAt != "" && strings.HasPrefix(dir, stopAt) && dir != stopAt {
+	if stopAt == "" {
+		return
+	}
+	if !strings.HasSuffix(stopAt, "/") {
+		stopAt += "/"
+	}
+	for strings.HasPrefix(dir, stopAt) && dir != stopAt {
 		err := os.Remove(dir)
 		if err != nil {
 			break
